requirements.txt
Django==4.1.7
asgiref==3.6.0
pip==22.3.1
setuptools==65.5.1
sqlparse==0.4.3
tzdata==2023.2
wheel==0.38.4
psycopg2==2.9.5
django-environ==0.10.0
djangorestframework==3.14.0
dj-rest-auth==3.0.0
djangorestframework-simplejwt==5.2.2